requirements.txt
aggdraw==1.3.18.post0
aspose-imaging-cloud==22.4.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
colormath==3.0.0
colorthief==0.2.1
comtypes==1.4.6
docopt==0.6.2
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.8
image-to-psd==1.1.0
imageio==2.35.1
joblib==1.4.2
lazy_loader==0.4
networkx==3.3
numpy==2.1.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
photoshop-python-api==0.22.4
PhotoshopAPI==0.4.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
psd-tools==1.9.34
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyDrive==1.3.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
tifffile==2024.8.10
uritemplate==4.1.1
urllib3==2.2.2
Wand==0.6.13
wheel==0.42.0
